+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export * from './actions';
 export * from './component';
-export * from './methods';
 export * from './reducer';
 export * from './utils';