+++ src/info/index.ts
@@ -1,3 +1,4 @@
 export * from './classInfo';
 export * from './componentInfo';
 export * from './creatorInfo';
+export * from './creatorMethods';