+++ test/info/creatorMethods.tests.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { component } from 'src';
-import { getCreatorMethods } from 'src/components';
+import { getCreatorMethods } from 'src/info';
 
 describe('component methods', () => {
     describe(nameof(getCreatorMethods), () => {