+++ src/utils/log.ts
@@ -22,12 +22,6 @@ class Log {
         console.warn('WARN [redux-app] ' + message, ...optionalParams);
     }
 
-    public error(message: string, ...optionalParams: any[]): void {
-        if (!this.shouldLog(LogLevel.Error))
-            return;
-        console.error('ERROR [redux-app] ' + message, ...optionalParams);
-    }
-
     private shouldLog(level: LogLevel): boolean {
         if (globalOptions.logLevel === LogLevel.None)
             return false;